add missing intents for SecurityScheme
