+++ lib/intent-models/oauth-flow.js
@@ -1,64 +1,65 @@
 module.exports = {};
 const IntentBase = require('./base');
-const { IntentOAuthFlow } = require('./oauth-flows');
 /**
  * 
  * @class
- * @alias module:@asyncapi/parser#IntentOAuthFlows
- * @extends IntentBase
+ * @alias module:@asyncapi/parser#IntentOAuthFlow
+ * @extends Base
  */
-class IntentOAuthFlows extends IntentBase {
+class IntentOAuthFlow extends IntentBase {
   /**
    * @returns {boolean}
    */
-  hasImplicit() {
+  hasAuthorizationUrl() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  implicit() {
-    return new IntentOAuthFlow();
+  authorizationUrl() {
+    return 'test.com';
   }
 
   /**
    * @returns {boolean}
    */
-  hasPassword() {
+  hasTokenUrl() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  password() {
-    return new IntentOAuthFlow();
+  tokenUrl() {
+    return 'test.com';
   }
 
   /**
    * @returns {boolean}
    */
-  hasClientCredentials() {
+  hasRefreshUrl() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  clientCredentials() {
-    return new IntentOAuthFlow();
+  refreshUrl() {
+    return 'test.com';
   }
 
   /**
    * @returns {boolean}
    */
-  hasAuthorizationCode() {
+  hasScopes() {
     return true; 
   }
   /**
-   * @returns {string|undefined}
+   * @returns {Object<string, string>|undefined}
    */
-  authorizationCode() {
-    return new IntentOAuthFlow();
+  scopes() {
+    return {
+      test: 'testscope'
+    };
   }
 }
 
-module.exports.IntentOAuthFlows = IntentOAuthFlows;
+module.exports.IntentOAuthFlow = IntentOAuthFlow;
