+++ lib/intent-models/oauth-flows.js
@@ -1,65 +1,64 @@
 module.exports = {};
 const IntentBase = require('./base');
+const { IntentOAuthFlow } = require('./oauth-flow');
 /**
  * 
  * @class
- * @alias module:@asyncapi/parser#IntentOAuthFlow
- * @extends Base
+ * @alias module:@asyncapi/parser#IntentOAuthFlows
+ * @extends IntentBase
  */
-class IntentOAuthFlow extends IntentBase {
+class IntentOAuthFlows extends IntentBase {
   /**
    * @returns {boolean}
    */
-  hasAuthorizationUrl() {
+  hasImplicit() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  authorizationUrl() {
-    return 'test.com';
+  implicit() {
+    return new IntentOAuthFlow();
   }
 
   /**
    * @returns {boolean}
    */
-  hasTokenUrl() {
+  hasPassword() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  tokenUrl() {
-    return 'test.com';
+  password() {
+    return new IntentOAuthFlow();
   }
 
   /**
    * @returns {boolean}
    */
-  hasRefreshUrl() {
+  hasClientCredentials() {
     return true; 
   }
   /**
    * @returns {string|undefined}
    */
-  refreshUrl() {
-    return 'test.com';
+  clientCredentials() {
+    return new IntentOAuthFlow();
   }
 
   /**
    * @returns {boolean}
    */
-  hasScopes() {
+  hasAuthorizationCode() {
     return true; 
   }
   /**
-   * @returns {Object<string, string>|undefined}
+   * @returns {string|undefined}
    */
-  scopes() {
-    return {
-      test: 'testscope'
-    };
+  authorizationCode() {
+    return new IntentOAuthFlow();
   }
 }
 
-module.exports.IntentOAuthFlow = IntentOAuthFlow;
+module.exports.IntentOAuthFlows = IntentOAuthFlows;
