+++ lib/intent-models/security-scheme.js
@@ -1,4 +1,6 @@
 const IntentBase = require('./Base');
+const IntentOAuthFlows = require('./oauth-flows')
+
 /**
  * @enum {string}
  */
@@ -24,43 +26,58 @@ const ApiKeyLocations = {
   HttpApiKeyHeader: 'header',
   HttpApiKeyCookie: 'cookie',
 };
+
 /**
  * 
  * @class
  * @alias module:@asyncapi/parser#IntentSecurityScheme
  * @extends IntentBase
- * @returns {SecurityScheme}
  */
 class IntentSecurityScheme extends IntentBase {
   /**
-     * @returns {Types}
-     */
-  type() { return Types.UserPassword;}
+   * @returns {Types}
+   */
+  type() { return Types.UserPassword; }
     
   /**
-     * @returns {string}
-     */
+   * @returns {string}
+   */
   name() { return null; }
     
   /**
-     * @returns {ApiKeyLocations}
-     */
+   * @returns {ApiKeyLocations}
+   */
   in() { return null; }
     
   /**
-     * @returns {string}
-     */
+   * @returns {string}
+   */
   scheme() { return null; }
     
   /**
-     * @returns {string}
-     */
+   * @returns {string}
+   */
   bearerFormat() { return null; }
-    
+
   /**
-     * @returns {string}
-     */
+   * @returns {boolean}
+   */
+  hasDescription() { return true; }
+  
+  /**
+   * @returns {string}
+   */
   description() { return 'Security Scheme description'; }
+
+  /**
+   * @returns {IntentOAuthFlows}
+   */
+  flows() { return new IntentOAuthFlows(); }
+
+  /**
+   * @returns {string}
+   */
+  openIdConnectUrl() { return 'https://asyncapi.com'; }
 }
 
 module.exports = IntentSecurityScheme;
